requirements.txt
altair==5.3.0
beautifulsoup4==4.12.3
matplotlib==3.9.1
pandas==2.2.2
Pillow==10.4.0
pypalettes==0.1.1
python-dotenv==1.0.1
Requests==2.32.3
squarify==0.4.4
st_social_media_links==0.1.3
streamlit==1.36.0
streamlit_extras==0.4.3
